+++ healvis/cosmology.py
@@ -2,8 +2,6 @@
 # Copyright (c) 2019 Radio Astronomy Software Group
 # Licensed under the 3-clause BSD License
 
-from __future__ import absolute_import, division, print_function
-
 import numpy as np
 from astropy.cosmology import Planck15 as cosmo
 from astropy.constants import c