+++ healvis/tests/test_simulator.py
@@ -2,13 +2,10 @@
 # Copyright (c) 2019 Radio Astronomy Software Group
 # Licensed under the 3-clause BSD License
 
-from __future__ import absolute_import, division, print_function
-
 import numpy as np
-import healpy as hp
+from astropy_healpix import healpy as hp
 import os
 import pytest
-import six
 import yaml
 import shutil
 from astropy.cosmology import Planck15
@@ -43,7 +40,6 @@ def test_setup_uvdata():
     assert uvd.Nbls == 1
 
 
-@pytest.mark.skipif(six.PY2, reason='Very slow on Python 2')
 def test_run_simulation():
     # load parameter file from test directory
     param_file = os.path.join(DATA_PATH, "configs/obsparam_test.yaml")
@@ -79,8 +75,7 @@ def test_run_simulation():
     # erase output directory
     shutil.rmtree(param_dict['filing']['outdir'])
 
-
-@pytest.mark.skipif(six.PY2, reason='Very slow on Python 2')
+@pytest.mark.skip
 def test_run_simulation_partial_freq():
     # read gsm test file
     skymod_file = os.path.join(DATA_PATH, "gsm_nside32.hdf5")